Fixed Last Move Errors:
Think the last move errors are gone. The problem was that my error
statements within playerTurn were while statements and not ifs. This
meant that when the game should have been over, the playerTurn called
within the while loop terminated, and then continued from that loop
within the second playerTurn loop. Now, the playerTurn method can end
only once when the game is finished.