Forces users to Pass:
In Othello, if a player has no possible moves, their turn is
automatically passed. This has been incorporated.

In addition, the number of moves displayed appears to be fixed. Where I
implemented the count++ impacted what was happening so I made a couple
of code adjustments. Not pretty, but it works.